1. Added code for reading properties, train and test files from the executable's directory. 2. Some experimentation with the Iris dataset.

+++ NeuralNetwork/NeuralNetwork.h
@@ -14,7 +14,7 @@ class NeuralNetwork
 	int activationFuncTypeOuputLayer;
 	double learningRate;
 	std::vector<Layer*> layers;
-	int target;
+	double target;
 public:
 	NeuralNetwork(std::string nnConfigFilename);
 	~NeuralNetwork();
